116:  simplify metadata
use default values from rootlayout in child layouts
